test(worker): move host validation tests into worker.test.ts
test/regression/issue/ is reserved for true regressions; this behavior
was never correct, so the tests belong with the module's coverage.
